vicodec_test: add the ability to save the encoded file with ioctl API
Add a `--save <output_file>` option that maps the CAPTURE buffers and
saves the encoded stream into a file. Since the approach for mapping
buffers is not decided yet for the device API, this is only effective
with the ioctl API.

Note: it would be better to pass a closure to the encoder functions that
takes the encoded data as argument, so `main.rs` is reponsible for
saving the file. This would remove duplicate code between both
implementations.