vicodec_pipeline: move output queue to client
Let the output queue be directly controlled by the client thread. This
removes the need to pass frames through a command, and lets the client
work directly with QBuffers to prepare the frames as it wants, while
also enabling it to map MMAP buffers - something that would not be
possible if they resided in another buffer.

As it turns out, this also makes the code shorter.