vicodec_pipeline: add the ability to wait for an available OUTPUT buffer
The client now manages the OUTPUT queue by itself, which means it must
also dequeue the buffers. We try to do this transparently, but doing so
puts us at risk of doing busy-loops until a buffer is available.

Address this by adding a new command asking the encoder thread to
also monitor OUTPUT buffers availability in its pool loop, and run a
callback when once becomes available. We can use the callback to signal
a barrier, and have the client thread wait on it. When it goes past the
barrier, it knows that an OUTPUT buffer must be available.