vicodec_pipeline: use encoder-thread side callback to process encoded buffers
Pass a callback that the encoder thread will call for each encoded
buffer it dequeues, instead of passing them through a message to the
client. This reduces latency, simplifies the code, and the client still
has the option to implement a message-passing mechanism in its own
callback if it needs the buffers anyway.