vicodec_pipeline/encoder: dequeue as many CAPTURE buffers as we can after a poll()
It seems like poll() won't return immediately if we don't dequeue all
the CAPTURE buffers that were available when it first returned. If we
call it again without dequeuing everything we can, it will only return
the next time a CAPTURE buffer becomes ready, which can cause deadlocks
at the end of streams.