device/dqbuf: put buffer handles behind an option
Buffer handles need to be moved from the DQBuf object, which implements
the Drop trait and thus needs to replace any content moved out of it, as
we do with its buffer handles. But not all handles implement Default -
notably DualBufferHandles don't. So put them behind an Option and add a
method that guarantees it will return the handles once, and only once.