memory: complete overhaul
Rewrite the memory model into something simpler and more flexible. In
particular this model will allow us to support dynamically choosing the
memory type of a queue in a much simpler way.

This is a deep rewrite so changes are all over the place. I was too lazy
to split properly. >_<