device/qbuf: overhaul of interface and internals
This is an almost-rewrite of the qbuf mechanism in device/. It separates
the QBuffer interface into its Capture and Output components, which have
fundamentally different interfaces, and thus makes it impossible to pass
bytes_used for CAPTURE buffers (which does not make sense to do).

Having this separation will allow us to enable generic CAPTURE buffers
in the encoder and decoder.

It also replaces the clumsy Plane struct by directly passing the
required parameters in the queue*() methods when they are needed.