ioctl/dqbuf: make interface safer and FFI-friendly
Make the interface to DQBuffer simpler and safer by embedding the
v4l2_buffer and just providing accessor methods to its data. DQBuffer is
read-only information so this way works pretty well.

Additionally, we can now get a pointer to the v4l2_buffer which is
convenient to pass to the C FFI if needed.