bindings: fix layout for 32-bit ioctls
The layout generated by bindgen includes padding members for unions.
These padding members are tailored for the host architecture, meaning
the bindings will be generated with 64-bit alignment in mind. This
causes a serious problem when building 32-bit binaries since the
structures used for the ioctls will then be misaligned.

Work around this issue for now by removing these members.