device: Poller overhaul
Rework the Poller class to allow the user to use as many Wakers as
desired. This makes it possible to distinguish from several non-device
event sources.

This apparently trivial change requires to completely change the way we
report events, so do that as well even if that makes the commit a bit
messy.