decoder/stateful: move capture buffer waker into relevant state
The capture buffer waker is only relevant in the decoding state, so
store it there. Moreover, this forces us to reallocate a new waker with
every resolution change, which is a good thing: if we use the same
waker, we may get a signal for older buffers that we are not interested
in anymore.