device: make GetFree*Buffer and Get*BufferByIndex return the same type
Until now the two ways of obtaining a buffer for queueing (by index, or
any available buffer) were defining their own return type. Even though
in practice that type is always the same, this fact wasn't obvious to
the type system and it was impossible to e.g. affect the return value of
either method to the same variable.

Fix this by definining new OutputQueueableProvider and
CaptureQueueableProvider traits with their own Queueable inner type,
that must be implemented by implementors of both methods. This
guarantees that they both return the same type to the type system.

Also this way of doing lets us remove some code that was specific to
primitive or generic buffer, so it's arguably a better design overall.