decoder/stateful: obtain visible rectangle and pass it to client
The size of the buffers used to decode frames may differ from the useful
area of the frame, due to aligment constraints enforced by hardware. To
mitigate this, V4L2 decoders provide the actual visible rectangle
through the COMPOSE selection target of the CAPTURE queue. Obtain and
pass this information.