decoder/stateful: replace exit waker with bool
Using a waker to break the loop is overkill when a simple boolean can do
the same thing. Moreover the waker forces us to enter the poll loop
after the last buffer of the stream has been dequeued, making the V4L2
device return POLLERR. The boolean approach prevents this issue.