ioctl/dqbuf: store plane data into Box
The v4l2_buffer instance contains a raw pointer to an array of
v4l2_planes which we can pass to the C ABI. This is effectively playing
with fire as the DQBuf instances can be moved around. Make sure that
pointer stays stable by moving the data it points to into a Box.