ioctl/querybuf: make conversion from v4l2_buffer a fallible operation
Make the conversion from v4l2_buffer capable of failure. This allows us
to validate the field when converting to a V4l2Buffer, which in turns
allows us to return safe types in its methods later, at the cost of an
added generic in the error types of operations that do the conversion.