lib: overhaul of extended controls
This (too big) change rethinks the way we are handling controls. It does
several things, which should each be their own CL but doing so is
unfortunately difficult:

* Define control structures as transparent wrappers of
  v4l2_ext_control, along with code to make them safe to manipulate,
* Allow g/s/_ext_ctrls to process several controls in one call, as they
  were intended to be,
* Spread control-related code around the codebase, to avoid having
  everything in the `ioctl` module.
* Document things, even if partially.

Although there is still some work to do in order to make controls easier
to manage, this new approach should be more in-line with how V4L2
expects controls to work.