lib: remove shady use of transmute
We assumed that transmute could safely be used between the m fields of
v4l2_plane and v4l2_buffer, but that actually depends on the size of a
pointer on the platform. Use a match statement using the memory type
since it is known in both situations, and to this effect to this effect
remove the trait method that was used in only one place anyway.

Fixes #24.