ioctl: revamp struct V4l2Buffer
Change the way struct V4l2Buffer works in order to (hopefully) make it
safer.

The invariants that this structure is supposed to maintain are now
better defined, and it is possible to obtain a valid pointer to the
inner v4l2_buffer with a valid planes pointer (if the buffer is
multiplanar).

Plane data is not accessed through queue-agnostic accessors, that
operate the same whether the buffer is single-planar or not.

Documentation for V4l2Buffer is better, albeit still incomplete.

There is still some work to be done, but this looks like a step in the
right direction.