lib: generate bindings during build
The pre-generated bindings make it easier to build the library, but also
introduce some limitations for workflows that require a specific version
of the headers to be used - e.g. Android wants to generate the bindings
from their own headers.

Address this by removing the runbindgen crate that we used to create our
pre-generated bindings, and replacing it with a (mostly similar)
build.rs file that generates them at build time from the videodev2.h
header. We also update the Android.bp build files to leverage a
custom-built generator that creates the bindings and makes them
available for the v4l2r crate to include.

This will likely introduce issues with older headers that do not define
symbols that the library used. These will be addressed via conditional
compilation as they are discovered.

Confirmed to build using `cargo` on a Linux system that has
`/usr/include/linux/videodev2.h`, and on AOSP when checked into
`external/rust/crates/v4l2r`.