device/queue: remove lifetime requirement on GetBufferByIndex trait
The goal of this trait was to ensure the obtained buffer holds a
reference to this queue of origin, but this is overstretched: that
constraint can be expressed in the implementations of the trait, and
having it in the trait definition prevents us from implementing more
dynamic mechanisms to borrow buffers from queues, i.e. using reference
counting and runtime checks instead of compile-time constraints.

Remove this constraint from the trait definitions ; the implementations
also express it, so they hold just as well without it.